# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Space Collab Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProjectById
      tags: [projects]
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{id}/join:
    post:
      operationId: joinProject
      tags: [projects]
      summary: Join a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{id}/leave:
    post:
      operationId: leaveProject
      tags: [projects]
      summary: Leave a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{id}/comments:
    get:
      operationId: listProjectComments
      tags: [projects]
      summary: List project comments/discussions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addProjectComment
      tags: [projects]
      summary: Add a comment to project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /projects/{id}/experiments:
    get:
      operationId: listProjectExperiments
      tags: [projects]
      summary: List project experiments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of experiments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experiment"
    post:
      operationId: addProjectExperiment
      tags: [projects]
      summary: Add experiment note/result
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExperimentRequest"
      responses:
        "201":
          description: Experiment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        grade:
          type: string
        role:
          type: string
          enum: [student, admin]
        bio:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - role
        - createdAt
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [planning, active, completed, archived]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        ownerId:
          type: integer
        ownerName:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        isMember:
          type: boolean
      required:
        - id
        - title
        - description
        - category
        - status
        - progress
        - ownerId
        - memberCount
        - createdAt
        - updatedAt
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/User"
    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - category
    UpdateProjectRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [planning, active, completed, archived]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        tags:
          type: array
          items:
            type: string
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        projectId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - projectId
        - userId
        - createdAt
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    Experiment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        results:
          type: string
        projectId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: string
        fileUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - projectId
        - userId
        - createdAt
    CreateExperimentRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        results:
          type: string
        fileUrl:
          type: string
      required:
        - title
        - description
...