# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Collab Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from "./project";
import type { User } from "./user";

export type ProjectDetail = Project & {
  members?: User[];
};